rekisteröintinäkymä refaktoroitu
